+++ js/data.js
@@ -1,3 +1,4 @@
+import {getRandomElement, getRandomIntInclusive} from './utils.js';
 const NAMES = [
   'Андрей',
   'Игорь',
@@ -34,4 +35,22 @@ const DESCRIPTION = [
 
 const PHOTO_NUMBER_ELEMENTS = 25;
 
-export {NAMES,MESSAGES,DESCRIPTION, PHOTO_NUMBER_ELEMENTS};
+//функция для создания photo
+const createPhoto = () => ({
+
+  id: getRandomIntInclusive(1, 25),
+  url: `photos/${getRandomIntInclusive(1, 25)}.jpg`,
+  description: getRandomElement(DESCRIPTION),
+  likes: getRandomIntInclusive(15, 200),
+  comments: {
+    id: getRandomIntInclusive(1, 200),
+    avatar:` img/avatar-${getRandomIntInclusive(0, 6)}.svg`,
+    message: getRandomElement(MESSAGES),
+    name:getRandomElement(NAMES),
+  },
+});
+
+
+//СОЗДАНИЕ МАССИВА ИЗ N ЭЛЕМЕНТОВ
+const photo = Array.from({length:PHOTO_NUMBER_ELEMENTS}, createPhoto);
+console.log(photo);
