+++ js/main.js
@@ -1,0 +1,2 @@
+import './data.js';
+
